MIM-57 拆分双 Runtime 分页回归测试
